extract Variable refactoring
moved live bus tracker styles to its own styles page
