+++ ConcordiaMaps/styles/LiveBusTrackerStyles.js
@@ -1,0 +1,23 @@
+import { StyleSheet } from "react-native";
+const styles = StyleSheet.create({
+  markerImage: {
+    width: 40,
+    height: 40,
+    borderRadius: 20,
+  },
+  errorContainer: {
+    position: "absolute",
+    bottom: 20,
+    left: 20,
+    right: 20,
+    backgroundColor: "rgba(255,0,0,0.8)",
+    padding: 10,
+    borderRadius: 5,
+  },
+  errorText: {
+    color: "white",
+    textAlign: "center",
+  },
+});
+
+export default styles;
